# openapi.yaml
openapi: 3.0.0
info:
  title: Filmoteka API
  description: Api task to VK Intern
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /actor:
    post:
      summary: Add actor information
      security:
        - AdminAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Actor'
        required: true
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        '400':
          description: Corrupted data in request body
        '500':
          description: Server error
    get:
      summary: Get actors list
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Actor"
        '500':
          description: Server error
  /actor/{actor_id}:
    put:
      summary: Refactor actor information
      security:
        - AdminAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Actor'
        required: true
      responses:
        '201':
          description: Refactored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        '400':
          description: Body data corrupted
        '404':
          description: Actor doesn't exists in the database
        '500':
          description: Server error
    delete:
      summary: Delete current actor from db
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Deleted successfully
        '400':
          description: Actor doesn't exists in db
        '404':
          description: Actor not found
  /film:
    post:
      summary: Add film information
      security:
        - AdminAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
        required: true
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        '400':
          description: Corrupted data in request body
        '500':
          description: Server error
    get:
      summary: Get films list
      parameters:
        - name: search
          in: query
          description: Part of film name
          schema:
            type: string
            default: ''
          required: false
        - name: actor
          in: query
          description: Part of actor name
          schema:
            type: string
            default: ''
          required: false
        - name: ordering
          in: query
          description: Answer ordering
          schema:
            type: string
            enum: ['asc', 'desc']
            default: 'desc'
          required: false
        - name: field
          in: query
          description: Field to order by
          schema:
            type: string
            enum: ['name', 'rating', 'date']
            default: 'rating'
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Film"
        '500':
          description: Server error
  /film/{film_id}:
    put:
      summary: Refactor film information
      security:
        - AdminAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
        required: true
      responses:
        '201':
          description: Refactored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
        '400':
          description: Body data corrupted
        '404':
          description: Film doesn't exist in the db
    delete:
      summary: Delete current film from db
      security:
        - AdminAuth: []
      responses:
        '201':
          description: Deleted successfully
        '400':
          description: Film doesn't exist in the db
        '404':
          description: User has no rights to delete

components:
  schemas:
    Actor:
      type: object
      properties:
        id:
          type: number
          required: false
          readOnly: true
          uniqueItems: true
        name:
          type: string
        gender:
          type: boolean
          description: true if man, false if woman
        birthday:
          type: string
          example: "2006-01-02"
        films:
          type: array
          items:
              $ref: '#/components/schemas/Film'
    Film:
      type: object
      properties:
        id:
          type: number
          required: false
          readOnly: true
          uniqueItems: true
        name:
          type: string
          minLength: 1
          maxLength: 150
        description:
          maxLength: 1000
        date:
          type: string
          example: "2006-01-02"
        rating:
          type: number
          minimum: 0
          maximum: 10
        Actors:
          type: array
          items:
            $ref: '#/components/schemas/Actor'


  securitySchemes:
    AdminAuth:
      type: server
      scheme: bearer
      bearerFormat: JWT